allow 0.5 offset in AMR volume testing
added comment explaining this situation. Copied here:

the internal data layout for AMR is cell-centered, but the data is
provided with the expectation of vertex-centered. This leads to a small
consistent offset. In the case of using the getZValue procedural function
this is an offset of 0.5 (i.e. objectCoordinates.z - 0.5). The exception
to this is when z == 0, in which case we are at data boundaries and the
value is 0. This occurs with all AMR interpolation methods