dynamic data in USM for unstructured volumes
Tell embree to use USM when it builds BVHs.

Note: since we are not using embree's rtcThreadLocalAllocator
we are fully responsible for deleting the allocated memory.
This includes the ispcrt level structs.
Approach is to have a vector of allocated items, and a mutex
so that embree callbacks construct the vector safely.

Next create unstructured volume temporary arrays in USM space.

This ensures that the prims put into embree BVH are USM safe
so that the GPU side iterator/sampler kernels can get to them.

Similarly, the parameter data arrays, for example geometry,
that application passes in also must be in USM and thus safe
for kernels on the GPU to access.